requirements.txt
contextily
fastparquet
geopandas
graphviz
jaxlib==0.3.22
jsonlines
jupyterlab
matplotlib
numpy
onnx
openpyxl
pydot
scikit-learn
seaborn
tensorflow-addons
tqdm
voila
wandb==0.13.10
